Update Core SDK and runtime from 2.1 to 3.1
Moved from netstandard2.0 to netcoreapp3.1 since this assembly
is very .NET Core dependent.

Any Microsoft.NET.Sdk projects need a FrameworkReference to
Microsoft.AspNetCore.App. If the project is Microsoft.NET.Sdk.Web,
then there is an implicit reference. This is necessary when using
a shared framework, as we do with Docker.

There were a number of NuGet packages in the Microsoft.AspNetCore
namespace that didn't have versions beyond 2.2.0. Those were
apparently incorporated into the meta package and so no longer need
to be referenced.