Fix issue with capture
InvalidOperationException: The client projection contains a reference to a
constant expression of 'Scullery.EntityFrameworkCore.EntityFrameworkJobStore'
through the instance method 'ToJobDescriptor'. This could potentially cause
a memory leak; consider making the method static so that it does not capture
constant in the instance. See https://go.microsoft.com/fwlink/?linkid=2103067
for more information.